requirements.txt
Flask
smbus2
RPi.GPIO
pyserial
